+++ test/field.campaign-code.spec.js
@@ -23,6 +23,16 @@ describe('field.campaign-code', () => {
 
   describe('type/source',  () => {
     logger.clear();
+    it('use guid', async() => {
+      let r = await f.convert('campaignCode',   { "_parent": "campaign", "code": "uitnodiging", guid: 'CC_TEST' });
+      assert.equal(r.typeId, '123', 'found default')
+      assert.equal(r.typeGuid, 'CC_TEST');
+      r = await f.convert('campaignCode',   { "_parent": "campaign", "code": "uitnodiging", typeGuid: 'CC_TEST' });
+      assert.equal(r.typeId, '123', 'found default')
+      assert.equal(r.typeGuid, 'CC_TEST');
+
+    });
+
     it('translate default', async () => {
       let r = await f.convert('campaigncode', {code: 'test', _parent: 'test'}, logger);
       assert.equal(r.typeId, 10, 'found default');
@@ -36,7 +46,9 @@ describe('field.campaign-code', () => {
     it('use type', async () => {
       let r = await f.convert('campaignCode',   { "_parent": "campaign", "code": "uitnodiging" });
       assert.equal(r.typeId, '123', 'found default')
-    })
+    });
+
+
   });
 
 });